all the code has been moved from NetCDFReader into ExperimentalData; code cleanup
